fix: rename core to misc
