+++ jaxley/modules/base.py
@@ -26,7 +26,7 @@ from jaxley.utils.cell_utils import (
     interpolate_xyz,
     loc_of_index,
 )
-from jaxley.utils.core_utils import childview, concat_and_ignore_empty
+from jaxley.utils.misc_utils import childview, concat_and_ignore_empty
 from jaxley.utils.plot_utils import plot_morph
